feature : ChatBot Refactoring - 챗봇 레이아웃 완성
