+++ src/components/chatbot/Chatbot.css
@@ -12,75 +12,19 @@
   height: 80vh;
 }
 
-.react-chatbot-kit-chat-header {
-}
-
-.react-chatbot-kit-chat-input-container {
-}
-
 .react-chatbot-kit-chat-message-container {
   /*실제로 메세지가 보이는 컨테이너*/
   height: 70vh;
 }
 
-.react-chatbot-kit-chat-input {
-}
-
-.react-chatbot-kit-chat-input-form {
+.react-chatbot-kit-chat-input ::before {
+  content: "hi";
 }
 
 .react-chatbot-kit-chat-input::placeholder {
-}
-
-.react-chatbot-kit-chat-btn-send {
-}
-
-.react-chatbot-kit-chat-btn-send-icon {
+  visibility: hidden;
 }
 
 .react-chatbot-kit-chat-bot-message-container {
   width: 300px;
-}
-
-.react-chatbot-kit-chat-bot-avatar-container {
-}
-
-.react-chatbot-kit-chat-bot-avatar-icon {
-}
-
-.react-chatbot-kit-chat-bot-avatar-letter {
-}
-
-.react-chatbot-kit-chat-bot-message {
-  width: 300px;
-}
-
-.react-chatbot-kit-chat-bot-message-arrow {
-}
-
-.react-chatbot-kit-chat-bot-loading-icon-container {
-}
-
-.chatbot-loader-container {
-}
-
-#chatbot-loader #chatbot-loader-dot1 {
-}
-
-#chatbot-loader #chatbot-loader-dot2 {
-}
-
-#chatbot-loader #chatbot-loader-dot3 {
-}
-
-.react-chatbot-kit-error {
-}
-
-.react-chatbot-kit-error-container {
-}
-
-.react-chatbot-kit-error-header {
-}
-
-.react-chatbot-kit-error-docs {
 }
