+++ src/components/chatbot/MessageParser.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
+import { createChatBotMessage } from "react-chatbot-kit";
 
-const MessageParser = ({ children, action1, action2 }) => {
+const MessageParser = ({ children, action1 }) => {
   const [items, setitems] = useState(0);
   const [chatState, setChatState] = useState(0);
   const functionKeys = Object.keys(action1);
@@ -21,8 +22,9 @@ const MessageParser = ({ children, action1, action2 }) => {
   }, [action1, chatState, functionKeys]);
 
   const parse = () => {
-    setitems(1);
+    setChatState(0);
     console.log(items);
+    createChatBotMessage("hi");
   };
 
   return (
@@ -31,7 +33,6 @@ const MessageParser = ({ children, action1, action2 }) => {
         return React.cloneElement(child, {
           parse: parse,
           action1,
-          action2,
         });
       })}
     </div>
