+++ src/constatnts/VocieFishingScenario.js
@@ -2,4 +2,6 @@ const scenario = [
   { sub: "bot", text: "hi" },
   { sub: "client", text: "client" },
   { sub: "bot", text: "hi2" },
+  { sub: "client", text: "굿이야" },
 ];
+export default scenario;
